refactor: move rotating in given direction from extension.ts to dispatchDirectionalRotation in project.ts, create DeviceRotationDirection enum
